requirements.txt
python-frontmatter
pandas
jupyter